+++ src/Map.js
@@ -21,6 +21,7 @@ import Spin from './Spin'
 import Alert from './Alert'
 import TileLoadingMonitor from './TileLoadingMonitor'
 import ButtonPanel from './ButtonPanel'
+import NearestTreeArrows from './NearestTreeArrows'
 
 class MapError extends Error {}
 
@@ -148,6 +149,7 @@ export default class Map {
     divContainer.style.height = '100%'
     divContainer.style.position = 'relative'
     divContainer.innerHTML = `
+      <div id="greenstand-nearest-tree-arrow" style="position: absolute; top: 0; left: 0; width: 100%; height: 100%"></div>
       <div id="greenstand-leaflet" style="position: relative;width: 100%;height: 100%;"></div>
       <div id="greenstand-map-spin" style="z-index: 999; position: absolute; width: 100%; top: 0px; left: 0px" ></div>
       <div id="greenstand-map-alert" style="z-index: 999; position: absolute; width: 100%; top: 0px; left: 0px" ></div>
@@ -177,10 +179,38 @@ export default class Map {
         () => this.goPrevPoint(),
       )
       this.buttonPanel.mount(mountButtonPanelTarget)
-      this.on(Map.REGISTERED_EVENTS.TREE_SELECTED, () =>
-        this.buttonPanel.show(),
-      )
+      this.on(Map.REGISTERED_EVENTS.TREE_SELECTED, () => {
+        const currentPoint = this.layerSelected.payload
+        const points = this.getPoints()
+        const index = points.reduce((a, c, i) => {
+          if (c.id === currentPoint.id) {
+            return i
+          }
+          return a
+        }, -1)
+        if (points.length <= 1) {
+          return null
+        }
+        this.buttonPanel.show()
+        if (index === 0) {
+          this.buttonPanel.hideLeftArrow()
+        } else if (index === points.length - 1) {
+          this.buttonPanel.hideRightArrow()
+        } else {
+          this.buttonPanel.showLeftArrow()
+          this.buttonPanel.showRightArrow()
+        }
+      })
     }
+
+    // Nearest Tree Arrow
+    const mountNearestArrowTarget = document.getElementById(
+      'greenstand-nearest-tree-arrow',
+    )
+    this.nearestTreeArrow = new NearestTreeArrows(() =>
+      this.moveToNearestTree(),
+    )
+    this.nearestTreeArrow.mount(mountNearestArrowTarget)
 
     // load google map
     await this.loadGoogleSatellite()
@@ -217,6 +247,7 @@ export default class Map {
       // mount event
       this.map.on('moveend', (e) => {
         log.warn('move end', e)
+        this.checkArrow()
         this.events.emit(Map.REGISTERED_EVENTS.MOVE_END)
       })
 
@@ -332,6 +363,17 @@ export default class Map {
         res()
       })
     })
+  }
+
+  async addGeoJson(source) {
+    let geo = source
+
+    if (typeof source === 'string') {
+      geo = (await axios.get(source)).data
+    }
+
+    const layer = window.L.geoJSON(geo).addTo(this.map)
+    return layer
   }
 
   async gotoBounds(bounds) {
@@ -469,7 +511,7 @@ export default class Map {
       const isLoading = this.layerUtfGrid.isLoading()
       log.warn('utf layer is loading:', isLoading)
       if (isLoading) {
-        log.error('can not handle the grid utf check when loading, cancel!')
+        log.warn('can not handle the grid utf check when loading, cancel!')
         return false
       }
       const begin = Date.now()
@@ -1105,12 +1147,26 @@ export default class Map {
       const nearest = await this.getNearest()
       if (nearest) {
         const placement = this.calculatePlacement(nearest)
-        if (this.onFindNearestAt) {
-          this.onFindNearestAt(placement)
-        }
+        this.handleNearestArrowDisplay(placement)
       } else {
         log.warn("Can't get the nearest:", nearest)
+        this.handleNearestArrowDisplay()
       }
+    }
+  }
+
+  handleNearestArrowDisplay(placement) {
+    !placement || placement === 'in'
+      ? this.nearestTreeArrow.hideArrow()
+      : this.nearestTreeArrow.showArrow(placement)
+  }
+
+  async moveToNearestTree() {
+    const nearest = await this.getNearest()
+    if (nearest) {
+      this.goto(nearest)
+    } else {
+      log.warn('can not find nearest:', nearest)
     }
   }